install: checking if the file exists before creating symlink might be better to check if the file is a symlink and if so replace it. And otherwise, if the file exist, do nothing

+++ install.sh
@@ -7,7 +7,7 @@ fi
 NAME="$(basename "$0")"
 
 for FILE in *.sh; do
-    if [ "$NAME" != "$FILE" ]; then
+    if [ "$NAME" != "$FILE" ] && [ ! -e "$HOME/.scripts/$FILE" ]; then
         ln -s "$PWD/$FILE" "$HOME/.scripts"
     fi
 done
